% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add2.R
\name{add2}
\alias{add2}
\title{add2}
\usage{
add2(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
The sum of \code{a} and \code{b}
}
\description{
Adds two numbers together.
}
\examples{
add2(1, 2)
}
