% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task9.R
\name{addthree}
\alias{addthree}
\title{Three value sum}
\usage{
addthree(x = 1, y = 2, z = 3)
}
\arguments{
\item{x, y, z}{are all values to enter}
}
\value{
this is the sum of the three values entered
}
\description{
This function sums three values
}
\examples{
addthree(x=2, y=4, z=6)
}
