% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_9_function.R
\name{Task_9_function}
\alias{Task_9_function}
\title{Task_9_function}
\usage{
Task_9_function(x, name)
}
\arguments{
\item{x}{a dataframe that has two columns, if it doesn't don't worry you'll know.}

\item{name}{the name of the third column, it can be whatever you want it to be.}
}
\description{
This function allows you to add two columns in a data frame.
}
\examples{

gooddataframe <- data.frame(
a=c(1,2,3,4,5,6,7),
b=c(1,2,3,4,5,6,7)
)

Task_9_function(gooddataframe)
}
\keyword{adding}
